% Auto-generated: do not edit by hand
\name{dashIfc}

\alias{dashIfc}

\title{DashIfc component}

\description{

}

\usage{
dashIfc(id=NULL, url=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify the container for the IFC viewer component.}

\item{url}{Character. The url for where the IFC file is hosted.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

